Moving auth to discoverable package
