+++ core/src/main/java/io/paradoxical/cassieq/admin/resources/api/v1/QueueDebugResource.java
@@ -1,7 +1,7 @@
 package io.paradoxical.cassieq.admin.resources.api.v1;
 
 import com.google.inject.Inject;
-import io.paradoxical.cassieq.auth.AccountAuth;
+import io.paradoxical.cassieq.discoverable.auth.AccountAuth;
 import io.paradoxical.cassieq.dataAccess.interfaces.MonotonicRepository;
 import io.paradoxical.cassieq.dataAccess.interfaces.PointerRepository;
 import io.paradoxical.cassieq.resources.api.BaseQueueResource;
